draft ingest and data
