+++ packages/azure-kusto-data/dist-esm/package.json
@@ -1,6 +1,6 @@
 {
     "name": "azure-kusto-data",
-    "version": "3.4.2",
+    "version": "3.4.3",
     "description": "Azure Data Explorer Query SDK",
     "main": "dist-esm/src/index.js",
     "module": "dist-esm/src/index.js",
@@ -20,14 +20,7 @@
         "http": false,
         "crypto": false,
         "stream": false,
-        "./dist-esm/src/tokenProvider.js": "./dist-esm/src/tokenProvider.browser.js",
-        "./dist-esm/src/tokenProvider.ts": "./dist-esm/src/tokenProvider.browser.ts"
-    },
-    "exports": {
-        ".": {
-            "node": "./src/index.js",
-            "browser": "./src/index.browser.js"
-        }
+        "./dist-esm/src/tokenProvider.js": "./dist-esm/src/tokenProvider.browser.js"
     },
     "engines": {
         "node": ">= 14.0.0"
@@ -60,10 +53,6 @@
     "author": "",
     "license": "ISC",
     "homepage": "https://github.com/Azure/azure-kusto-node/blob/master/azure-kusto-data/README.md",
-    "optionalDependencies": {
-        "@azure/msal-browser": "^2.28.1",
-        "@azure/msal-node": "^1.12.1"
-    },
     "devDependencies": {
         "@types/mocha": "^9.1.1",
         "@types/webpack-dev-server": "^4.7.2",
