+++ packages/azure-kusto-ingest/package.json
@@ -1,9 +1,10 @@
 {
     "name": "azure-kusto-ingest",
-    "version": "3.4.2",
+    "version": "3.4.3",
     "description": "Azure Data Explorer Ingestion SDK",
-    "main": "index.js",
-    "types": "index.d.ts",
+    "main": "dist-esm/src/index.js",
+    "module": "dist-esm/src/index.js",
+	"types" : "dist-esm/index.d.ts",
     "scripts": {
         "build": "tsc"
     },
@@ -29,15 +30,27 @@
         "directory": "azure-kusto-ingest"
     },
     "files": [
-        "source/**/*.js",
-        "source/**/*.d.ts",
-        "index.js",
-        "index.d.ts",
+        "dist/",
+        "dist-esm/src/",
+        "dist-esm/package.json",
         "tsconfig.tsbuildinfo",
         "example.js"
     ],
     "author": "",
     "license": "ISC",
+    "browser": {
+        "fs": false,
+        "os": false,
+        "process": false,
+        "stream-http": false,
+        "https": false,
+        "http": false,
+        "crypto" : false,
+        "stream": "readable-stream",
+        "./dist-esm/src/streamUtils.js": "./dist-esm/src/streamUtils.browser.js",
+        "./dist-esm/src/fileDescriptor.js": "./dist-esm/src/fileDescriptor.browser.js",
+        "./dist-esm/src/ingestClient.js": "./dist-esm/src/ingestClient.browser.ts"
+    },
     "dependencies": {
         "@azure/storage-blob": "^12.11.0",
         "@azure/storage-queue": "^12.10.0",
@@ -47,8 +60,9 @@
         "@types/stream-to-array": "^2.3.0",
         "@types/uuid": "^8.3.4",
         "@types/uuid-validate": "0.0.1",
-        "azure-kusto-data": "file:../azure-kusto-data",
+        "azure-kusto-data": "^3.4.3",
         "moment": "^2.29.4",
+        "readable-stream": "~2.1.0",
         "stream-array": "^1.1.2",
         "stream-to-array": "^2.3.0",
         "tmp-promise": "^3.0.3",
