+++ packages/azure-kusto-ingest/src/columnMappings.ts
