+++ packages/azure-kusto-ingest/src/errors.ts
