+++ packages/azure-kusto-ingest/src/ingestionBlobInfo.ts
