+++ packages/azure-kusto-ingest/src/ingestionProperties.ts
