+++ packages/azure-kusto-ingest/src/status.ts
