+++ packages/azure-kusto-ingest/src/statusQ.ts
