+++ packages/azure-kusto-ingest/src/streamUtils.browser.ts
@@ -1,17 +1,42 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { CompressionType, FileDescriptor, StreamDescriptor } from "./descriptors";
-import fs from "fs";
+import { CompressionType, StreamDescriptor } from "./descriptors";
+import { FileDescriptor } from "./fileDescriptor";
 import { PassThrough, Readable } from "stream";
 import streamify from "stream-array";
 
-export const fileToStream = (file: FileDescriptor | string): StreamDescriptor => {
+export const fileToStream = async (file: Blob): Promise<StreamDescriptor> => {
     const fileDescriptor = file instanceof FileDescriptor ? file : new FileDescriptor(file);
-    const streamFs = fs.createReadStream(fileDescriptor.filePath);
+    const streamFs = file.stream();
     const compressionType = fileDescriptor.zipped ? CompressionType.GZIP : CompressionType.None;
     return new StreamDescriptor(streamFs, fileDescriptor.sourceId, compressionType);
 };
+
+export const fileToBuffer = async (file: Blob): Promise<StreamDescriptor> => {
+    try {
+        const fileDescriptor = file instanceof FileDescriptor ? file : new FileDescriptor(file);
+        const buffer = await streamToBuffer(file.stream());
+        const compressionType = fileDescriptor.zipped ? CompressionType.GZIP : CompressionType.None;
+        return new StreamDescriptor(buffer, fileDescriptor.sourceId, compressionType);
+    } catch(error) {
+        return await fileToStream(file);
+    }
+}
+
+// A helper method used to read a Node.js readable stream into a Buffer
+async function streamToBuffer(readableStream: NodeJS.ReadableStream): Promise<Buffer> {
+    return new Promise((resolve, reject) => {
+      const chunks: Buffer[] = [];
+      readableStream.on("data", (data: Buffer | string) => {
+        chunks.push(data instanceof Buffer ? data : Buffer.from(data));
+      });
+      readableStream.on("end", () => {
+        resolve(Buffer.concat(chunks));
+      });
+      readableStream.on("error", reject);
+    });
+  }
 
 export const mergeStreams = (...streams: Readable[]): Readable => {
     let pass = new PassThrough();
