+++ packages/azure-kusto-ingest/src/streamingIngestClient.ts
@@ -3,7 +3,8 @@
 
 import { IngestionPropertiesInput } from "./ingestionProperties";
 
-import { CompressionType, FileDescriptor, StreamDescriptor } from "./descriptors";
+import { CompressionType, StreamDescriptor } from "./descriptors";
+import { FileDescriptor } from "./fileDescriptor";
 import zlib from "zlib";
 import { AbstractKustoClient } from "./abstractKustoClient";
 import { Client as KustoClient, KustoConnectionStringBuilder } from "azure-kusto-data";
@@ -24,8 +25,10 @@ class KustoStreamingIngestClient extends AbstractKustoClient {
         const props = this._getMergedProps(ingestionProperties);
         const descriptor: StreamDescriptor = stream instanceof StreamDescriptor ? stream : new StreamDescriptor(stream);
 
-        const compressedStream = descriptor.compressionType === CompressionType.None ? descriptor.stream.pipe(zlib.createGzip()) : descriptor.stream;
-        return this.kustoClient.executeStreamingIngest(
+        const compressedStream = descriptor.compressionType === CompressionType.None
+            ? (!(descriptor.stream instanceof Buffer) ? descriptor.stream.pipe(zlib.createGzip()) : descriptor.stream)
+            : descriptor.stream;
+        return await this.kustoClient.executeStreamingIngest(
             props.database as string,
             props.table as string,
             compressedStream,
@@ -36,7 +39,7 @@ class KustoStreamingIngestClient extends AbstractKustoClient {
     }
 
     async ingestFromFile(file: FileDescriptor | string, ingestionProperties?: IngestionPropertiesInput): Promise<KustoResponseDataSet> {
-        return this.ingestFromStream(fileToStream(file), ingestionProperties);
+        return this.ingestFromStream(await fileToStream(file), ingestionProperties);
     }
 }
