+++ packages/azure-kusto-ingest/test/e2eTests/e2eTest.ts
@@ -5,16 +5,16 @@
 
 import assert from "assert";
 import fs, { ReadStream } from "fs";
-import IngestClient from "../../source/ingestClient";
-import KustoIngestStatusQueues from "../../source/status";
+import IngestClient from "../../src/ingestClient";
+import KustoIngestStatusQueues from "../../src/status";
 import { Client, ClientRequestProperties, KustoConnectionStringBuilder as ConnectionStringBuilder } from "azure-kusto-data/dist-esm/src";
-import StreamingIngestClient from "../../source/streamingIngestClient";
-import ManagedStreamingIngestClient from "../../source/managedStreamingIngestClient";
-import { CompressionType, StreamDescriptor } from "../../source/descriptors";
-import { DataFormat, IngestionProperties, ReportLevel } from "../../source/ingestionProperties";
+import StreamingIngestClient from "../../src/streamingIngestClient";
+import ManagedStreamingIngestClient from "../../src/managedStreamingIngestClient";
+import { CompressionType, StreamDescriptor } from "../../src/descriptors";
+import { DataFormat, IngestionProperties, ReportLevel } from "../../src/ingestionProperties";
 import { CloudSettings } from "azure-kusto-data/src/cloudSettings";
-import { sleep } from "../../source/retry";
-import { JsonColumnMapping } from "../../source/columnMappings";
+import { sleep } from "../../src/retry";
+import { JsonColumnMapping } from "../../src/columnMappings";
 
 interface ParsedJsonMapping {
     Properties: { Path: string };
