+++ packages/azure-kusto-ingest/test/ingestClientTest.ts
@@ -2,11 +2,11 @@
 // Licensed under the MIT License.
 
 import assert from "assert";
-import { KustoIngestClient } from "../source/ingestClient";
-import { DataFormat, IngestionProperties, IngestionPropertiesInput, ReportLevel, ReportMethod } from "../source/ingestionProperties";
-import { IngestionPropertiesValidationError } from "../source/errors";
-import KustoStreamingIngestClient from "../source/streamingIngestClient";
-import KustoManagedStreamingIngestClient from "../source/managedStreamingIngestClient";
+import { KustoIngestClient } from "../src/ingestClient";
+import { DataFormat, IngestionProperties, IngestionPropertiesInput, ReportLevel, ReportMethod } from "../src/ingestionProperties";
+import { IngestionPropertiesValidationError } from "../src/errors";
+import KustoStreamingIngestClient from "../src/streamingIngestClient";
+import KustoManagedStreamingIngestClient from "../src/managedStreamingIngestClient";
 import { KustoConnectionStringBuilder } from "azure-kusto-data";
 
 describe("KustoIngestClient", () => {
