+++ packages/azure-kusto-ingest/tsconfig.json
@@ -1,5 +1,12 @@
 {
     "extends": "../../tsconfig.json",
-    "include": ["./source", "./test", "index.ts"],
+    "include": ["src/**/*.ts", "test/**/*.ts","webpack.config.ts"],
+    "compilerOptions": {
+        "outDir": "./dist-esm",
+        "paths": {
+          "azure-kusto-ingest": ["./src/index"]
+        },
+        "lib": ["DOM"]
+    },
     "exclude": ["node_modules", "*.d.ts", "**/*.d.ts"]
 }
