+++ packages/quick_start/README.md
@@ -1,7 +1,7 @@
 # Quickstart App
 
 The quickstart application is a **self-contained and runnable** example app that demonstrates authenticating, connecting to, administering, ingesting data
-into and querying Azure Data Explorer using the azure-kusto-node SDK.
+into and querying Azure Data Explorer using the azure-kusto-node SDK. // TODO chang when ready to azure-kusto-js here and repo name in github!
 You can use it as a baseline to write your own first kusto client application, altering the code as you go, or copy code sections out of it into your app.
 
 **Tip:** The app includes comments with tips on recommendations, coding best practices, links to reference materials and recommended TODO changes when adapting the code to your needs.
