+++ packages/quick_start/package.json
@@ -9,8 +9,8 @@
         "quick_start": "npx ts-node source/SampleApp.ts"
     },
     "dependencies": {
-        "azure-kusto-data": "^3.4.2",
-        "azure-kusto-ingest": "^3.4.2",
+        "azure-kusto-data": "^3.4.3",
+        "azure-kusto-ingest": "^3.4.3",
         "diff": "^5.1.0",
         "ts-node": "^10.9.1",
         "uuid": "^8.3.2"
