+++ packages/quick_start/source/Utils.ts
@@ -120,12 +120,12 @@ export class Authentication extends Utils {
         certificatePath: string | undefined,
         certificatePassword: string | undefined,
         applicationId: string | undefined,
-        tenantId: string | undefined
+        tenantId: string | undefined,
+        sendX5c?: boolean
     ): Promise<KustoConnectionStringBuilder> {
         const appId: string | undefined = process.env.APP_ID;
         const appTenant: string | undefined = process.env.APP_TENANT;
         const privateKeyPemFilePath: string | undefined = process.env.PRIVATE_KEY_PEM_FILE_PATH;
-        const certThumbprint: string | undefined = process.env.CERT_THUMBPRINT;
         const publicCertFilePath: string | undefined = process.env.PUBLIC_CERT_FILE_PATH;
 
         if (!certificatePath || !certificatePassword || !applicationId || !tenantId || !appId) {
@@ -136,26 +136,21 @@ export class Authentication extends Utils {
             } else {
                 const pemCertificate: string = await fs.promises.readFile(privateKeyPemFilePath, "utf8");
                 if (publicCertFilePath) {
-                    const publicCertificate: string = await fs.promises.readFile(publicCertFilePath, "utf8");
                     return KustoConnectionStringBuilder.withAadApplicationCertificateAuthentication(
                         clusterUri,
                         appId,
                         pemCertificate,
-                        publicCertificate,
-                        appTenant
+                        appTenant,
+                        sendX5c
                     );
                 }
-                if (!certThumbprint) {
-                    this.errorHandler(`"Missing required field: "certThumbprint" in environment in order to authenticate using a certificate."`);
-                } else {
-                    return KustoConnectionStringBuilder.withAadApplicationCertificateAuthentication(
-                        clusterUri,
-                        appId,
-                        pemCertificate,
-                        certThumbprint,
-                        appTenant
-                    );
-                }
+                return KustoConnectionStringBuilder.withAadApplicationCertificateAuthentication(
+                    clusterUri,
+                    appId,
+                    pemCertificate,
+                    appTenant
+                );
+
             }
         }
     }
